% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataframe.R, R/write_dataframe.R
\name{read_dataframe}
\alias{read_dataframe}
\alias{write_dataframe}
\title{write and read data frames to plain text}
\usage{
read_dataframe(), write_dataframe()

read_dataframe(), write_dataframe()
}
\arguments{
\item{one}{dataframe or a path}

\item{one}{dataframe or a path}
}
\value{
returns a data frame readed from files

returns a data frame readed from files
}
\description{
write and read data frames to plain text

write and read data frames to plain text
}
