% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_factor.R
\name{detect_factor}
\alias{detect_factor}
\title{Factors detect whether the input is a character}
\usage{
detect_factor()
}
\arguments{
\item{the}{parameter you want to detect}
}
\value{
"input is a character"/ "Input is a factor"
}
\description{
If the input is a character, print "input is a character"  
If the input is a factor, print "Input is a factor"
}
\examples{
detect_factor(factor(c("a", "a", "b"))) "Input is a factor"
}
