% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_factor.R
\name{reorder_factor}
\alias{reorder_factor}
\title{Reorder a factor in descending order}
\usage{
factor_reorder(x)
}
\arguments{
\item{x}{a factor you want to reorder}
}
\value{
Returns the input factor in descending order
An error is raised if the input is not a factor
}
\description{
a version of reorder() that uses desc() a la (d)plyr
}
\examples{
reorder_factor(factor(c("B", "A", "D"))) # Levels: D B A
}
