% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_factor.R
\name{set_factor}
\alias{set_factor}
\title{Set a factor in a given order}
\usage{
set_factor(x)
}
\arguments{
\item{x}{a factor you want to set}
}
\value{
Return levels to the order in which they appear in the data,
An error is raised if the input is not a factor
}
\description{
sets levels to the order in which they appear in the data, i.e. set the levels “as is”
}
\examples{
set_factor(factor(c("b", "c", "a"))) # Levels: b c a
}
